Highlight and Row Additions
Worked on adding the Share Name and Customer Name cells into the output dataframe.
Worked on the highlight (background) coloring of cells. Having issues with finding Bold options. Need to play with Openxlsx package styling. Once have a demo need to try with R Shiny to make sure that will work in unison.