Order Only Bug and Week/Date Addition
FS team found a bug in the Order Only members where I did not have the correct text and I hadn't accounted for new rank fields so was never filtering members when I should have been.
Added Week # and Date to form which needs to be manually entered each week for each file.