Started on App
Started on the first page of the Shiny App. Start with the date selection and file upload. I still have three more actions to set into motion to then be used later on. Also need to find a way to use icons/colors to say certain steps are complete. The new files were created to help with the test process to make sure errors are handled accordingly.