Flaw in Column Optimization
The FS team found a flaw where I was missing members in the output. I found that I wasn't accounting for when Column 2 filled up faster than Column 1. Once this was accounted for this seemed to remedy the issue. Also put in an error check to make sure an error is thrown when the row count is exceeded.